ia64: Fix gru compiler warning
Fix compiler warning in sgi-gru driver.

+++ drivers/misc/sgi-gru/gru_instructions.h
@@ -29,7 +29,7 @@
 #if defined(CONFIG_IA64)
 #include <linux/compiler.h>
 #include <asm/intrinsics.h>
-#define __flush_cache(p)		ia64_fc(p)
+#define __flush_cache(p)		ia64_fc((unsigned long)p)
 /* Use volatile on IA64 to ensure ordering via st4.rel */
 #define gru_ordered_store_int(p,v)					\
 		do {							\
